Python/requirements.txt
python-telegram-bot==13.7
requests==2.32.3
redis==4.5.5